+++ src/Chart/RomanNumberTest.php
@@ -25,14 +25,14 @@ final class RomanNumberTest extends TestCase
      *
      * @param array<string> $expected
      */
-    public function testGetLabels(int $nbLabels, int $label, Casing $lettercase, array $expected): void
+    public function testGetLabels(int $nbLabels, int $label, LetterCase $lettercase, array $expected): void
     {
         $generator = new RomanNumber(new DecimalNumber($label), $lettercase);
         self::assertSame($expected, iterator_to_array($generator->generate($nbLabels), false));
     }
 
     /**
-     * @return iterable<string, array{nbLabels:int, label:int, lettercase:Casing, expected:array<string>}>
+     * @return iterable<string, array{nbLabels:int, label:int, lettercase:LetterCase, expected:array<string>}>
      */
     public function providerLetter(): iterable
     {
@@ -40,63 +40,42 @@ final class RomanNumberTest extends TestCase
             'empty labels' => [
                 'nbLabels' => 0,
                 'label' => 1,
-                'lettercase' => Casing::UPPER,
+                'lettercase' => LetterCase::UPPER,
                 'expected' => [],
             ],
             'labels starts at 3' => [
                 'nbLabels' => 1,
                 'label' => 3,
-                'lettercase' => Casing::UPPER,
+                'lettercase' => LetterCase::UPPER,
                 'expected' => ['III'],
             ],
             'labels starts ends at 4' => [
                 'nbLabels' => 2,
                 'label' => 4,
-                'lettercase' => Casing::UPPER,
+                'lettercase' => LetterCase::UPPER,
                 'expected' => ['IV', 'V'],
             ],
             'labels starts at 0 (1)' => [
                 'nbLabels' => 1,
                 'label' => -1,
-                'lettercase' => Casing::LOWER,
+                'lettercase' => LetterCase::LOWER,
                 'expected' => ['i'],
             ],
             'labels starts at 0 (2)' => [
                 'nbLabels' => 1,
                 'label' => 0,
-                'lettercase' => Casing::LOWER,
+                'lettercase' => LetterCase::LOWER,
                 'expected' => ['i'],
             ],
         ];
     }
 
-    public function testStartWith(): void
-    {
-        $generator = new RomanNumber(new DecimalNumber(42));
-        self::assertSame(42, $generator->decimalNumber->startingAt);
-        $new = $generator->startsWith(69);
-        self::assertNotSame($new, $generator);
-        self::assertSame(69, $new->decimalNumber->startingAt);
-        self::assertSame($generator, $generator->startsWith(42));
-        self::assertSame(1, (new DecimalNumber(-3))->startingAt);
-        self::assertSame(1, $generator->startsWith(-3)->decimalNumber->startingAt);
-    }
-
-    public function testLetterCase(): void
-    {
-        $generator = new RomanNumber(new DecimalNumber(1));
-        self::assertTrue($generator->isUpper());
-        $new = $generator->case(Casing::LOWER);
-        self::assertFalse($new->isUpper());
-        $alt = $new->case(Casing::LOWER);
-        self::assertSame($alt, $new);
-    }
-
     public function testFormat(): void
     {
-        $generator = new RomanNumber(new DecimalNumber(10));
-        $newGenerator = $generator->case(Casing::LOWER);
-        self::assertSame('FOOBAR', $generator->format('fOoBaR'));
-        self::assertSame('foobar', $newGenerator->format('fOoBaR'));
+        $upperRoman = new RomanNumber(new DecimalNumber(10), LetterCase::UPPER);
+        $lowerRoman = new RomanNumber(new DecimalNumber(10), LetterCase::LOWER);
+
+        self::assertSame('FOOBAR', $upperRoman->format('fOoBaR'));
+        self::assertSame('foobar', $lowerRoman->format('fOoBaR'));
     }
 }